FIX: remove double periods (#42)
Anytime that an author/editor had a middle initial and is the sole author/editor, the template produces a sequence of double periods. This fix adds a `replace` helper to Handlebars, and uses it to automatically replace a sequence of `..` with `.`. A known issue with this is that it will replace *any* instance of `..` in the citation, which in some rare cases may be intentional.

closes #42