FIX: reverse editors when no author is present (#43)
When there is no author present, the editors are listed first, in place of the authors. In this case, the first and last name of the first editor in the list need to be reversed. This fix checks whether the reference has an authors field, and reverses the first editor in the list if so.

closes #43